refactor: extract Credential, App, DNS, Cluster services from bridge.go (Phase 2)
Agent-Task: bridge-refactor-phase2
Agent-Model: SOLO
Agent-Decision: Extract 4 domain service files to reduce bridge.go from 2659 to 2094 lines

Extracted services:
- credential_service.go (164 lines): CreateCredential, ListCredentials, DeleteCredential,
  UpdateCredential, CreateCredentialWithExpiry, RotateCredential
- app_service.go (173 lines): ListApps, CreateApp, DeleteApp, GetAppDetail, UpdateApp,
  GetAppDeploymentHistory, SearchAppLogs
- dns_service.go (199 lines): getDNSProvider, DNSCreateRecord, DNSDeleteRecord,
  DNSListRecords, UpdateDNSRecord, BatchDNS
- cluster_service.go (90 lines): CreateCluster, GetCluster, ListClusters, UpdateCluster,
  DeleteCluster, TestClusterConnection

All methods remain on Bridge struct receiver to satisfy mcp.Deployer interface.
Removed unused dns import from bridge.go.
