+++ internal/service/bridge.go
@@ -27,7 +27,6 @@ import (
 	"github.com/Yogdunana/deploypilot/internal/monitor"
 	"github.com/Yogdunana/deploypilot/internal/plugin"
 	"github.com/Yogdunana/deploypilot/internal/provider/cicd"
-	"github.com/Yogdunana/deploypilot/internal/provider/dns"
 	registry "github.com/Yogdunana/deploypilot/internal/provider/registry"
 	"github.com/Yogdunana/deploypilot/internal/provider/server"
 	"github.com/Yogdunana/deploypilot/internal/sandbox"
@@ -173,41 +172,6 @@ func (b *Bridge) GetConfirmationStore() *confirm.Store {
 }
 
 // getDNSProvider loads the first enabled DNS provider from DB and returns a DNSProvider interface.
-func (b *Bridge) getDNSProvider(ctx context.Context) (dns.DNSProvider, error) {
-	if b.DB == nil {
-		return nil, fmt.Errorf("database not available")
-	}
-	var provider model.Provider
-	err := b.DB.Where("type LIKE ? AND enabled = ?", "dns-%", true).First(&provider).Error
-	if err != nil {
-		return nil, fmt.Errorf("no enabled DNS provider found: %w", err)
-	}
-	// Parse config JSON
-	var cfg struct {
-		APIToken       string `json:"api_token"`
-		AccountEmail   string `json:"account_email"`
-		AccessKeyID    string `json:"access_key_id"`
-		AccessKeySecret string `json:"access_key_secret"`
-		SecretID       string `json:"secret_id"`
-		SecretKey      string `json:"secret_key"`
-	}
-	if err := json.Unmarshal([]byte(provider.Config), &cfg); err != nil {
-		return nil, fmt.Errorf("failed to parse DNS provider config: %w", err)
-	}
-
-	switch provider.Type {
-	case "dns-cloudflare":
-		return dns.NewCloudflareProvider(cfg.APIToken, cfg.AccountEmail), nil
-	case "dns-aliyun":
-		return dns.NewAliyunProvider(cfg.AccessKeyID, cfg.AccessKeySecret), nil
-	case "dns-tencent":
-		return dns.NewTencentProvider(cfg.SecretID, cfg.SecretKey), nil
-	case "dns-west-dns":
-		return dns.NewWestDNSProvider(cfg.APIToken, cfg.AccessKeySecret), nil
-	default:
-		return nil, fmt.Errorf("unsupported DNS provider type: %s", provider.Type)
-	}
-}
 
 // getRemoteExecutor creates an SSH executor for the given server.
 // It looks up the server record, finds its credential, decrypts the password/key,
@@ -680,32 +644,6 @@ func (b *Bridge) GetContainerStatus(ctx context.Context, name string) (*mcp.Cont
 
 // ---------- 3. ListApps ----------
 
-func (b *Bridge) ListApps(ctx context.Context) ([]mcp.ContainerStatus, error) {
-	var rows []map[string]interface{}
-	if err := b.DB.Table("apps").Find(&rows).Error; err != nil {
-		return nil, fmt.Errorf("failed to query apps: %w", err)
-	}
-
-	result := make([]mcp.ContainerStatus, 0, len(rows))
-	for _, r := range rows {
-		cs := mcp.ContainerStatus{
-			ID:     toString(r["id"]),
-			Name:   toString(r["name"]),
-			Image:  toString(r["container_name"]),
-			Status: toString(r["status"]),
-		}
-		if v, ok := r["env_vars"]; ok {
-			if s, ok := v.(string); ok && s != "" {
-				var m map[string]string
-				if json.Unmarshal([]byte(s), &m) == nil {
-					cs.Labels = m
-				}
-			}
-		}
-		result = append(result, cs)
-	}
-	return result, nil
-}
 
 // ---------- 4. ListServers ----------
 
@@ -731,49 +669,6 @@ func (b *Bridge) ListServers(ctx context.Context) ([]mcp.ServerInfo, error) {
 	return result, nil
 }
 
-// ---------- 5. CreateApp ----------
-
-func (b *Bridge) CreateApp(ctx context.Context, cfg mcp.CreateAppConfig) (string, error) {
-	id := uuid.New().String()
-	if err := b.DB.Table("apps").Create(map[string]interface{}{
-		"id":          id,
-		"name":        cfg.Name,
-		"repo_url":    cfg.RepoURL,
-		"branch":      defaultVal(cfg.Branch, "main"),
-		"domain":      cfg.Domain,
-		"tech_stack":  defaultVal(cfg.TechStack, "docker"),
-		"deploy_mode": defaultVal(cfg.DeployMode, "api"),
-		"server_id":   cfg.ServerID,
-		"status":      "created",
-	}).Error; err != nil {
-		return "", fmt.Errorf("failed to create app: %w", err)
-	}
-	return id, nil
-}
-
-// ---------- 6. DeleteApp ----------
-
-func (b *Bridge) DeleteApp(ctx context.Context, appID string) error {
-	row := make(map[string]interface{})
-	if err := b.DB.Table("apps").Where("id = ?", appID).Take(&row).Error; err != nil {
-		return fmt.Errorf("app not found: %w", err)
-	}
-
-	containerName := toString(row["container_name"])
-	if containerName != "" {
-		if err := b.d().Stop(ctx, containerName); err != nil {
-			slog.WarnContext(ctx, "failed to stop container during app deletion", "container", containerName, "error", err)
-		}
-		if err := b.d().Remove(ctx, containerName); err != nil {
-			slog.WarnContext(ctx, "failed to remove container during app deletion", "container", containerName, "error", err)
-		}
-	}
-
-	if err := b.DB.Table("apps").Where("id = ?", appID).Delete(nil).Error; err != nil {
-		return fmt.Errorf("failed to delete app from DB: %w", err)
-	}
-	return nil
-}
 
 // ---------- 7. Stop ----------
 
@@ -1030,24 +925,6 @@ func (b *Bridge) GetDeploymentHistory(ctx context.Context, containerName string,
 }
 
 // GetAppDeploymentHistory returns the deployment history for an app by app ID.
-func (b *Bridge) GetAppDeploymentHistory(ctx context.Context, appID string, limit int) ([]model.DeploymentRecord, error) {
-	if b.DB == nil {
-		return nil, fmt.Errorf("database not available")
-	}
-	if limit <= 0 || limit > 100 {
-		limit = 20
-	}
-
-	var records []model.DeploymentRecord
-	err := b.DB.Where("app_id = ?", appID).
-		Order("created_at DESC").
-		Limit(limit).
-		Find(&records).Error
-	if err != nil {
-		return nil, fmt.Errorf("failed to query app deployment history: %w", err)
-	}
-	return records, nil
-}
 
 // ---------- 11. DetectEnv ----------
 
@@ -1233,344 +1110,6 @@ func (b *Bridge) TestServer(ctx context.Context, serverID string) (interface{}, 
 
 // ---------- 16. CreateCredential ----------
 
-func (b *Bridge) CreateCredential(ctx context.Context, tenantID, name, credType, plainValue string) (interface{}, error) {
-	id := uuid.New().String()
-
-	// Encrypt the value before storing
-	encrypted, err := crypto.Encrypt(b.EncryptionKey, plainValue)
-	if err != nil {
-		return nil, fmt.Errorf("failed to encrypt credential: %w", err)
-	}
-
-	if err := b.DB.Table("credentials").Create(map[string]interface{}{
-		"id":              id,
-		"tenant_id":       tenantID,
-		"name":            name,
-		"type":            credType,
-		"encrypted_value": encrypted,
-	}).Error; err != nil {
-		return nil, fmt.Errorf("failed to create credential: %w", err)
-	}
-
-	return map[string]interface{}{
-		"id":        id,
-		"tenant_id": tenantID,
-		"name":      name,
-		"type":      credType,
-	}, nil
-}
-
-// ---------- 17. ListCredentials ----------
-
-func (b *Bridge) ListCredentials(ctx context.Context, tenantID string) (interface{}, error) {
-	var rows []map[string]interface{}
-	if err := b.DB.Table("credentials").Where("tenant_id = ?", tenantID).Find(&rows).Error; err != nil {
-		return nil, fmt.Errorf("failed to list credentials: %w", err)
-	}
-
-	// Mask values before returning
-	sanitized := make([]map[string]interface{}, 0, len(rows))
-	for _, r := range rows {
-		entry := map[string]interface{}{
-			"id":        toString(r["id"]),
-			"tenant_id": toString(r["tenant_id"]),
-			"name":      toString(r["name"]),
-			"type":      toString(r["type"]),
-		}
-		sanitized = append(sanitized, entry)
-	}
-	return sanitized, nil
-}
-
-// ---------- 18. DeleteCredential ----------
-
-func (b *Bridge) DeleteCredential(ctx context.Context, credID string) error {
-	result := b.DB.Table("credentials").Where("id = ?", credID).Delete(nil)
-	if result.Error != nil {
-		return fmt.Errorf("failed to delete credential: %w", result.Error)
-	}
-	if result.RowsAffected == 0 {
-		return fmt.Errorf("credential %s not found", credID)
-	}
-	return nil
-}
-
-// ---------- 19. DNSCreateRecord ----------
-
-func (b *Bridge) DNSCreateRecord(ctx context.Context, domain, recordType, name, value string) (interface{}, error) {
-	provider, err := b.getDNSProvider(ctx)
-	if err != nil {
-		slog.Error("DNS provider error", "error", err)
-		return map[string]interface{}{
-			"status":  "error",
-			"domain":  domain,
-			"type":    recordType,
-			"name":    name,
-			"value":   value,
-			"message": err.Error(),
-		}, nil
-	}
-	record := &dns.DNSRecord{
-		Domain: domain,
-		Type:   recordType,
-		Name:   name,
-		Value:  value,
-		TTL:    1,
-	}
-	if err := provider.CreateRecord(ctx, record); err != nil {
-		return map[string]interface{}{
-			"status":  "error",
-			"message": err.Error(),
-		}, nil
-	}
-	return map[string]interface{}{
-		"status": "success",
-		"domain": domain,
-		"type":   recordType,
-		"name":   name,
-		"value":  value,
-	}, nil
-}
-
-// ---------- 20. DNSDeleteRecord ----------
-
-func (b *Bridge) DNSDeleteRecord(ctx context.Context, recordID string) error {
-	provider, err := b.getDNSProvider(ctx)
-	if err != nil {
-		return err
-	}
-	// recordID format: "domain:type:name"
-	parts := strings.SplitN(recordID, ":", 3)
-	if len(parts) != 3 {
-		return fmt.Errorf("invalid record ID format, expected domain:type:name")
-	}
-	return provider.DeleteRecord(ctx, parts[0], parts[1], parts[2])
-}
-
-// ---------- 21. DNSListRecords ----------
-
-func (b *Bridge) DNSListRecords(ctx context.Context, domain string) (interface{}, error) {
-	provider, err := b.getDNSProvider(ctx)
-	if err != nil {
-		return map[string]interface{}{
-			"status":  "error",
-			"domain":  domain,
-			"message": err.Error(),
-		}, nil
-	}
-	records, err := provider.ListRecords(ctx, domain)
-	if err != nil {
-		return map[string]interface{}{
-			"status":  "error",
-			"message": err.Error(),
-		}, nil
-	}
-	// Convert to response format
-	result := make([]map[string]interface{}, 0, len(records))
-	for _, r := range records {
-		result = append(result, map[string]interface{}{
-			"domain":  r.Domain,
-			"type":    r.Type,
-			"name":    r.Name,
-			"value":   r.Value,
-			"ttl":     r.TTL,
-			"proxied": r.Proxied,
-		})
-	}
-	return map[string]interface{}{
-		"status":  "success",
-		"domain":  domain,
-		"records": result,
-	}, nil
-}
-
-// ---------- 25. GetAppDetail ----------
-
-func (b *Bridge) GetAppDetail(ctx context.Context, appID string) (interface{}, error) {
-	row := make(map[string]interface{})
-	if err := b.DB.Table("apps").Where("id = ?", appID).Take(&row).Error; err != nil {
-		return nil, fmt.Errorf("app not found: %w", err)
-	}
-	return row, nil
-}
-
-// ---------- 26. UpdateApp ----------
-
-func (b *Bridge) UpdateApp(ctx context.Context, appID string, config map[string]interface{}) (interface{}, error) {
-	if err := b.DB.Table("apps").Where("id = ?", appID).Updates(config).Error; err != nil {
-		return nil, fmt.Errorf("failed to update app: %w", err)
-	}
-
-	row := make(map[string]interface{})
-	if err := b.DB.Table("apps").Where("id = ?", appID).Take(&row).Error; err != nil {
-		return map[string]interface{}{"status": "updated", "id": appID}, nil
-	}
-	return row, nil
-}
-
-// ---------- 29. SearchAppLogs ----------
-
-func (b *Bridge) SearchAppLogs(ctx context.Context, appID, keyword string, limit int) (interface{}, error) {
-	// Look up container name from app record
-	row := make(map[string]interface{})
-	if err := b.DB.Table("apps").Where("id = ?", appID).Take(&row).Error; err != nil {
-		return nil, fmt.Errorf("app not found: %w", err)
-	}
-
-	containerName := toString(row["container_name"])
-	if containerName == "" {
-		containerName = toString(row["name"])
-	}
-
-	logs, err := b.d().GetContainerLogs(ctx, containerName, 2000)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get logs: %w", err)
-	}
-
-	// Filter lines containing the keyword
-	var matches []string
-	lines := strings.Split(logs, "\n")
-	for _, line := range lines {
-		if strings.Contains(line, keyword) {
-			matches = append(matches, line)
-			if len(matches) >= limit {
-				break
-			}
-		}
-	}
-
-	return map[string]interface{}{
-		"app_id":         appID,
-		"container":      containerName,
-		"keyword":        keyword,
-		"total_lines":    len(lines),
-		"match_count":    len(matches),
-		"limit":          limit,
-		"matching_lines": matches,
-	}, nil
-}
-
-// ---------- 30. UpdateDNSRecord ----------
-
-func (b *Bridge) UpdateDNSRecord(ctx context.Context, domain, subdomain, recordType, newValue string) (interface{}, error) {
-	provider, err := b.getDNSProvider(ctx)
-	if err != nil {
-		return map[string]interface{}{
-			"status":  "error",
-			"message": err.Error(),
-		}, nil
-	}
-	record := &dns.DNSRecord{
-		Domain: domain,
-		Type:   recordType,
-		Name:   subdomain,
-		Value:  newValue,
-		TTL:    1,
-	}
-	if err := provider.UpdateRecord(ctx, record); err != nil {
-		return map[string]interface{}{
-			"status":  "error",
-			"message": err.Error(),
-		}, nil
-	}
-	return map[string]interface{}{
-		"status":    "success",
-		"domain":    domain,
-		"subdomain": subdomain,
-		"type":      recordType,
-		"value":     newValue,
-	}, nil
-}
-
-// ---------- 31. UpdateCredential ----------
-
-func (b *Bridge) UpdateCredential(ctx context.Context, credID string, value string) (interface{}, error) {
-	// Encrypt the new value before storing
-	encrypted, err := crypto.Encrypt(b.EncryptionKey, value)
-	if err != nil {
-		return nil, fmt.Errorf("failed to encrypt credential: %w", err)
-	}
-
-	if err := b.DB.Table("credentials").Where("id = ?", credID).Update("encrypted_value", encrypted).Error; err != nil {
-		return nil, fmt.Errorf("failed to update credential: %w", err)
-	}
-	return map[string]interface{}{
-		"id":      credID,
-		"status":  "updated",
-		"message": "credential value updated",
-	}, nil
-}
-
-// ---------- 31b. CreateCredentialWithExpiry ----------
-
-func (b *Bridge) CreateCredentialWithExpiry(ctx context.Context, tenantID, name, credType, plainValue string, expiresInDays int) (interface{}, error) {
-	id := uuid.New().String()
-
-	// Encrypt the value before storing
-	encrypted, err := crypto.Encrypt(b.EncryptionKey, plainValue)
-	if err != nil {
-		return nil, fmt.Errorf("failed to encrypt credential: %w", err)
-	}
-
-	createData := map[string]interface{}{
-		"id":              id,
-		"tenant_id":       tenantID,
-		"name":            name,
-		"type":            credType,
-		"encrypted_value": encrypted,
-		"rotation_days":   expiresInDays,
-	}
-
-	// Set expires_at if expiresInDays > 0
-	if expiresInDays > 0 {
-		expiresAt := time.Now().Add(time.Duration(expiresInDays) * 24 * time.Hour)
-		createData["expires_at"] = expiresAt
-	}
-
-	if err := b.DB.Table("credentials").Create(createData).Error; err != nil {
-		return nil, fmt.Errorf("failed to create credential: %w", err)
-	}
-
-	result := map[string]interface{}{
-		"id":            id,
-		"tenant_id":     tenantID,
-		"name":          name,
-		"type":          credType,
-		"rotation_days": expiresInDays,
-	}
-	if expiresInDays > 0 {
-		result["expires_at"] = time.Now().Add(time.Duration(expiresInDays) * 24 * time.Hour).Format(time.RFC3339)
-	}
-	return result, nil
-}
-
-// ---------- 31c. RotateCredential ----------
-
-func (b *Bridge) RotateCredential(ctx context.Context, credID string, newPlainValue string) (interface{}, error) {
-	// Check credential exists
-	var row map[string]interface{}
-	if err := b.DB.Table("credentials").Where("id = ?", credID).Take(&row).Error; err != nil {
-		return nil, fmt.Errorf("credential not found: %w", err)
-	}
-
-	// Encrypt the new value
-	encrypted, err := crypto.Encrypt(b.EncryptionKey, newPlainValue)
-	if err != nil {
-		return nil, fmt.Errorf("failed to encrypt credential: %w", err)
-	}
-
-	now := time.Now()
-	updates := map[string]interface{}{
-		"encrypted_value": encrypted,
-		"last_rotated":    now,
-	}
-
-	if err := b.DB.Table("credentials").Where("id = ?", credID).Updates(updates).Error; err != nil {
-		return nil, fmt.Errorf("failed to rotate credential: %w", err)
-	}
-
-	return toString(row["name"]), nil
-}
 
 // ---------- 32. UpdateServer ----------
 
@@ -1792,32 +1331,6 @@ func (b *Bridge) BatchBackup(ctx context.Context, appIDs []string) (interface{},
 
 // ---------- 38. BatchDNS ----------
 
-func (b *Bridge) BatchDNS(ctx context.Context, records []map[string]interface{}) (interface{}, error) {
-	results := make([]map[string]interface{}, 0, len(records))
-	for i, rec := range records {
-		domain := toStringOrDefault(rec["domain"], "")
-		subdomain := toStringOrDefault(rec["subdomain"], "")
-		recordType := toStringOrDefault(rec["type"], "")
-		value := toStringOrDefault(rec["value"], "")
-
-		res, err := b.DNSCreateRecord(ctx, domain, recordType, subdomain, value)
-		status := "success"
-		if err != nil {
-			status = "error"
-		} else if m, ok := res.(map[string]interface{}); ok && m["status"] == "error" {
-			status = "error"
-		}
-		results = append(results, map[string]interface{}{
-			"index":  i,
-			"status": status,
-			"record": res,
-		})
-	}
-	return map[string]interface{}{
-		"total":   len(records),
-		"results": results,
-	}, nil
-}
 
 // ---------- 39. CheckSystemUpdate ----------
 // Implemented in upgrade.go
@@ -2081,85 +1594,7 @@ func (b *Bridge) TriggerCIBuild(ctx context.Context, providerType, repo, branch 
 // ========== Kubernetes Cluster Management ==========
 
 // CreateCluster creates a new Kubernetes cluster.
-func (b *Bridge) CreateCluster(ctx context.Context, cluster *model.Cluster) (*model.Cluster, error) {
-	if b.DB == nil {
-		return nil, fmt.Errorf("database not available")
-	}
-	return model.CreateCluster(cluster)
-}
 
-// GetCluster retrieves a cluster by ID.
-func (b *Bridge) GetCluster(ctx context.Context, id string) (*model.Cluster, error) {
-	if b.DB == nil {
-		return nil, fmt.Errorf("database not available")
-	}
-	return model.GetCluster(id)
-}
-
-// ListClusters returns all clusters for a tenant.
-func (b *Bridge) ListClusters(ctx context.Context, tenantID string) ([]model.Cluster, error) {
-	if b.DB == nil {
-		return nil, fmt.Errorf("database not available")
-	}
-	return model.ListClusters(tenantID)
-}
-
-// UpdateCluster updates a cluster's fields.
-func (b *Bridge) UpdateCluster(ctx context.Context, id string, updates map[string]interface{}) (*model.Cluster, error) {
-	if b.DB == nil {
-		return nil, fmt.Errorf("database not available")
-	}
-	return model.UpdateCluster(id, updates)
-}
-
-// DeleteCluster removes a cluster by ID.
-func (b *Bridge) DeleteCluster(ctx context.Context, id string) error {
-	if b.DB == nil {
-		return fmt.Errorf("database not available")
-	}
-	return model.DeleteCluster(id)
-}
-
-// TestClusterConnection tests connectivity to a Kubernetes cluster.
-func (b *Bridge) TestClusterConnection(ctx context.Context, id string) (interface{}, error) {
-	if b.DB == nil {
-		return nil, fmt.Errorf("database not available")
-	}
-
-	cluster, err := model.GetClusterWithSecrets(id)
-	if err != nil {
-		return nil, fmt.Errorf("cluster not found: %w", err)
-	}
-
-	k8sProvider, err := server.NewK8sProvider(cluster)
-	if err != nil {
-		return map[string]interface{}{
-			"status":  "error",
-			"message": fmt.Sprintf("failed to create k8s provider: %v", err),
-		}, nil
-	}
-
-	if err := k8sProvider.TestConnection(ctx); err != nil {
-		return map[string]interface{}{
-			"status":  "error",
-			"message": fmt.Sprintf("connection failed: %v", err),
-		}, nil
-	}
-
-	info, err := k8sProvider.GetClusterInfo(ctx)
-	if err != nil {
-		return map[string]interface{}{
-			"status":  "success",
-			"message": "connected",
-		}, nil
-	}
-
-	return map[string]interface{}{
-		"status": "success",
-		"message": "connected",
-		"info":   info,
-	}, nil
-}
 
 // ========== Kubernetes Deployment Operations ==========
